requirements.txt
torch>=1.8
torchvision
numpy
opencv-python
Pillow
